requirements.txt
requests
proton-driver
pycaret
python-dotenv